requirements.txt
# Core dependencies
confluent-kafka==2.3.0
websocket-client==1.6.4
pyyaml==6.0.1
python-dotenv==1.0.0

# Data processing
pandas>=2.2.2
pyarrow>=17.0.0
numpy>=1.24.0

# MLflow
mlflow>=2.15.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0  # For SMOTE and class balancing

# Monitoring and reporting
evidently>=0.4.16  # 0.4.16+ supports pydantic>=2.0.0

# Logging and utilities
structlog==23.2.0

# Jupyter for notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0

# API and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
prometheus-client>=0.19.0
httpx>=0.25.0  # For load testing

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

